requirement.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///C:/ci/aiohttp_1637857274009/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.10.8
cffi @ file:///C:/ci/cffi_1636523946399/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
cryptography @ file:///C:/ci/cryptography_1633520531101/work
cycler==0.11.0
flatbuffers @ file:///tmp/build/80754af9/python-flatbuffers_1614345733764/work
fonttools==4.29.0
frozenlist @ file:///C:/ci/frozenlist_1637767271796/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///tmp/build/80754af9/google-auth_1626320605116/work
google-auth-oauthlib==0.4.1
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///C:/ci/grpcio_1637590978642/work
h5py @ file:///C:/ci/h5py_1637120894255/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1638543082309/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.3.2
Markdown @ file:///C:/ci/markdown_1614364082838/work
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multidict @ file:///C:/ci/multidict_1607349747897/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1634088717532/work
oauthlib @ file:///tmp/build/80754af9/oauthlib_1623060228408/work
opencv-python==4.5.5.62
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging==21.3
Pillow==9.0.0
protobuf==3.14.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///C:/ci/pyjwt_1619682721924/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==3.0.7
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///C:/ci/scipy_1641555170412/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/ci/tensorboard-data-server_1633035225228/work/tensorboard_data_server-0.6.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
tensorflow==2.6.0
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor==1.1.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
Werkzeug @ file:///tmp/build/80754af9/werkzeug_1635505089296/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
wrapt @ file:///C:/ci/wrapt_1638434068867/work
yarl @ file:///C:/ci/yarl_1606940155993/work
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
